#449 moving delete_manager into its own module
