+++ medicines/doc-index-updater/src/service_bus_client/mod.rs
@@ -1,7 +1,5 @@
 use azure_sdk_core::errors::AzureError;
 use azure_sdk_service_bus::prelude::*;
-use std::time::Duration;
-use tokio::time::delay_for;
 
 pub async fn delete_queue_client() -> Result<Client, AzureError> {
     let service_bus_namespace = std::env::var("SERVICE_BUS_NAMESPACE")
@@ -33,17 +31,4 @@ pub async fn create_queue_client() -> Result<Client, AzureError> {
         .expect("Set env variable CREATE_QUEUE_POLICY_KEY first!");
 
     Client::new(service_bus_namespace, queue_name, policy_name, policy_key)
-}
-
-pub async fn delete_service_worker() -> Result<String, AzureError> {
-    let mut delete_client = delete_queue_client().await?;
-
-    loop {
-        let message = delete_client.peek_lock(time::Duration::days(1)).await?;
-        println!("{:?} message receive!", message);
-        // TODO: delete file in blob storage
-        // TODO: Update index
-        // TODO: Notify state manager
-        delay_for(Duration::from_secs(10)).await;
-    }
 }
